Fix a warning in the cast of Renderer9::mDeviceCaps.MaxAnisotropy.
See http://code.google.com/p/angleproject/source/detail?r=2000

TRAC #22826

Author: Jamie Madill

git-svn-id: https://angleproject.googlecode.com/svn/branches/es3proto@2105 736b8ea6-26fd-11df-bfd4-992fa37f6226

+++ src/libGLESv2/renderer/Renderer9.cpp
@@ -2251,7 +2251,7 @@ float Renderer9::getTextureMaxAnisotropy() const
 {
     if (mSupportsTextureFilterAnisotropy)
     {
-        return mDeviceCaps.MaxAnisotropy;
+        return static_cast<float>(mDeviceCaps.MaxAnisotropy);
     }
     return 1.0f;
 }
